Cleaning up compile warnings

+++ src/prompt.h
@@ -5,5 +5,5 @@
 int doeval();
 void print_stack();
 void print_eol();
-int doprompt(char *rx_buff, int rx_len, char *tx_buff, int tx_len);
+int doprompt(char *rx_buff, size_t rx_len, char *tx_buff, size_t tx_len);
